# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FormalAP1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/drivers": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "description": "Return all driver with their informations.",
                "operationId": "getAllDriver",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/drivers/{driver}": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "description": "Return driver by ID.",
                "operationId": "getAllDriver",
                "parameters": [
                    {
                        "$ref": "driver"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Drivers"
                ],
                "description": "Update driver by Id.",
                "operationId": "UpdateDriver",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Drivers"
                ],
                "description": "Update driver by Id.",
                "operationId": "UpdateDriver",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/drivers/": {
            "post": {
                "tags": [
                    "Drivers"
                ],
                "description": "Create new driver.",
                "operationId": "CreateDriver",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/races": {
            "get": {
                "tags": [
                    "Races"
                ],
                "description": "Return all Race with their informations.",
                "operationId": "getAllRace",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/races/{race}": {
            "get": {
                "tags": [
                    "Races"
                ],
                "description": "Return race by ID.",
                "operationId": "getAllrace",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Races"
                ],
                "description": "Update race by Id.",
                "operationId": "Updaterace",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Races"
                ],
                "description": "Update race by Id.",
                "operationId": "Updaterace",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/races/": {
            "post": {
                "tags": [
                    "Races"
                ],
                "description": "Create new Race.",
                "operationId": "CreateRace",
                "responses": {
                    "200": {
                        "description": "Opération réussi"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Interdit"
                    },
                    "400": {
                        "description": "Mauvaise requête"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "constants": {
                "L5_SWAGGER_CONST_HOST": "http://127.0.0.1:8000/api"
            }
        }
    ]
}